# Make Barthez tokenizer tests a bit faster