# What does this PR do?

Currently, CI is pretty slow because of this:
```
93.44s call     tests/test_tokenization_barthez.py::BarthezTokenizationTest::test_add_special_tokens
77.20s call     tests/test_tokenization_barthez.py::BarthezTokenizationTest::test_pretokenized_inputs
77.00s call     tests/test_tokenization_barthez.py::BarthezTokenizationTest::test_maximum_encoding_length_single_input
76.66s call     tests/test_tokenization_barthez.py::BarthezTokenizationTest::test_maximum_encoding_length_pair_input
75.77s call     tests/test_tokenization_barthez.py::BarthezTokenizationTest::test_internal_consistency
```

This is cause by the BarthezTokenizer conversion from slow to fast being pretty slow, so this PR saves the fast tokenizer to make those tests faster. To be even more efficient a new sentencepiece model with a mask token and a pad token should be added and used here.